draft(module-debug): write a lil more
i'll probably drop this or at least massively rewrite it, since after sleeping
on it (and working on tests for `<packages.nix/modules>.hostrr`), it feels like
the "building and configuring a VM" part isn't really the most important thing
for this scenario, but instead writing tests with `pkgs.nixosTest` is a lot
more ergonomic (but also has a few problems or gotchas). but the vm were the
things i was excited about and wanted to talk about when i started writing, so
now i don't know what to do :(

(though while writing those tests, all i wanted was a "definitive" guide to
nixpkgs test, because having to read 4 or 5 different random articles, pieces
of doc, and nixpkgs code just to understand what a python method did or why
something kept failing or what a attribute did was uuuh not great. i don't
know if i know enough or am readlly working on stuff other people would be
interested in, but i'd really like to see more in-depth nixos tests ressources
out there, and if i can help a little bit i'd be very proud :)